The nodes generation function has been updated
The nodes generation has been updated for better error handling. Also, now the function takes input arguments to check for solvability as well as save the variables generated. The function has also incorporated the changes made for movement of the blank tile, and now the function does not terminate after generating 100,000 nodes. Instead, it finds all the nodes possible till it finds a solution.